Extract editor command runners
